==> notebooks/2.1 Merge data [tax bill, land bank, postal vacancy].ipynb <==
# Combine Data (not from prop)

Data has now been filtered for the parcel numbers in the existing dataset. The end goal is to have a dataset with one row for each parcel in which to build the model. **Imputation will occur in a separate notebook!**

However, we can't forget to explore the existing data for trends or insight along the way, as well as engineer features as we see fit.

import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd
import numpy as np
import os
import datetime as dt

%matplotlib inline
sns.set_context('notebook')

path = 'Z:/largetransfer/luc/carter'

tci = pd.read_csv(path+'/data/model_data/tci_1_0.csv',dtype={'Parcel ID':str,'parcel':str}, parse_dates=['Date'])
tci = tci[['parcel','Date','vacant']]

ppns = set(tci['parcel'])
dates = dict(zip(tci.parcel, tci.Date))

print(tci.shape)


## County Land Bank
Creates and new variable in tht `tci_1_0.csv` which encodes parcels that have acquired as **1** and **0** based on information in the *land and bank* data set

lb = pd.read_csv(path+'/data/clean_data/county_lb.csv')

lb.info()

lb_ppns = set(lb[lb.disp_status=='Acquired'].parcel)

def get_lb(x, lb_ppns):
    if x in lb_ppns:
        return 1
    else:
        return 0

tci['lb_acquired'] = tci['parcel'].apply(lambda x: get_lb(x, lb_ppns))
tci['lb_tax_fc'] = tci['parcel'].apply(lambda x: get_lb(x,set(lb[lb.p_source=='Tax Foreclosure'].parcel)))

## Tax bill
Converts 
1. "TOTAL_NET_DELQ_BALANCE"
2. "GRAND_TOTAL_BALANCE"
3. "GRAND_TOTAL_OWED"
in both `taxbill_sep14` and `taxbill_may15.csv` into floating point numbers.

Next:
4. Creates a column,"delq_total_ratio"  for proportion of TOTAL_NET_DELQ_BALANCE
5. Creates a column, "paid_percent" for proportion of TOTAL_NET_DELQ_BALANCE
**NOTE:** all the above was that done for `taxbill_sep14` and `taxbill_may15.csv` respectively

**Finally:** specific set of columns indicated in "cols" was selected and merged with `tci_1_0.csv`

tb2 = pd.read_csv(path+'/data/clean_data/taxbill_sep14.csv', dtype=object)
tb = pd.read_csv(path+'/data/clean_data/taxbill_may15.csv', dtype=object)

tb['TOTAL_NET_DELQ_BALANCE']=tb['TOTAL_NET_DELQ_BALANCE'].apply(float)
tb2['TOTAL_NET_DELQ_BALANCE']=tb2['TOTAL_NET_DELQ_BALANCE'].apply(float)

tb['GRAND_TOTAL_BALANCE']=tb['GRAND_TOTAL_BALANCE'].apply(float)
tb2['GRAND_TOTAL_BALANCE']=tb2['GRAND_TOTAL_BALANCE'].apply(float)

tb['GRAND_TOTAL_OWED']=tb['GRAND_TOTAL_OWED'].apply(float)
tb2['GRAND_TOTAL_OWED']=tb2['GRAND_TOTAL_OWED'].apply(float)

cols = ['TOTAL_NET_DELQ_BALANCE','TAX_ASSESSED_LAND','TAX_MARKET_LAND',\
        'LENDER_PROCESS_TYPE','GRAND_TOTAL_BALANCE','GRAND_TOTAL_PAID','GRAND_TOTAL_OWED']

# tb.GRAND_TOTAL_BALANCE = tb.GRAND_TOTAL_BALANCE.apply(float)
# tb2.GRAND_TOTAL_BALANCE = tb2.GRAND_TOTAL_BALANCE.apply(float)

# tb.TOTAL_NET_DELQ_BALANCE = tb.TOTAL_NET_DELQ_BALANCE.apply(float)
# tb2.TOTAL_NET_DELQ_BALANCE = tb2.TOTAL_NET_DELQ_BALANCE.apply(float)

tb['delq_total_ratio'] = tb['TOTAL_NET_DELQ_BALANCE']/tb['GRAND_TOTAL_OWED']
tb['delq_total_ratio14'] = tb2['TOTAL_NET_DELQ_BALANCE']/tb2['GRAND_TOTAL_OWED']
tb['delq_total_ratio'] = tb['delq_total_ratio'].fillna(0)

tb['paid_percent'] = tb.GRAND_TOTAL_PAID.apply(float)/tb.GRAND_TOTAL_OWED.apply(float)
tb['paid_percent'] = tb['paid_percent'].fillna(1)

# tem = tci[tci.parcel.isin(tb[(tb['delq_total_ratio']>0.1)&(tb['delq_total_ratio14']>0.1)].PROPERTY_NUMBER)].vacant
# sum(tem),len(tem)

tb = tb.fillna(0)
tb2 = tb2.fillna(0)

tb = tb.set_index('PROPERTY_NUMBER')
tb2 = tb2.set_index('PROPERTY_NUMBER')

cols = ['TOTAL_NET_DELQ_BALANCE','TAX_ASSESSED_LAND','TAX_MARKET_LAND',\
        'LENDER_PROCESS_TYPE','GRAND_TOTAL_BALANCE','GRAND_TOTAL_PAID','delq_total_ratio','paid_percent']


tci = pd.merge(tci, tb[cols], left_on='parcel', right_index=True,how='left')
tci.loc[tci.LENDER_PROCESS_TYPE==0,'LENDER_PROCESS_TYPE'] = 'None'
tb2 = tb2.rename(columns={'GRAND_TOTAL_BALANCE':'GRAND_TOTAL_BALANCE_14'})
tci = pd.merge(tci, tb2[['GRAND_TOTAL_BALANCE_14']], left_on='parcel', right_index=True,how='left')

del tb, tb2

## Postal vacancy
1. Collects data from postal vacancy and added it to the `tci_1_0.csv`, excluding any duplicated observations
2. Finally saves it to `tci_2_1.csv`

pv = pd.read_csv(path+'/data/clean_data/postal_vacancy.csv', parse_dates=True, index_col=0)
pv.date = pv.date.apply(lambda x: dt.datetime(int(x[0:4]),int(x[5:7]),int(x[8:10])))
pv = pv.sort_values('date',ascending=False)

pv.columns

dates = list()
tp = list()
tn = list()
fp = list()
fn = list()
for time in sorted(list(set(pv.date))):
    dates.append(time)
    pv2 = pd.merge(tci, pv[(pv.date==time)], how='left', left_on='parcel', right_on='PARCEL')
    pv2['vindall'] = pv2['vindall'].fillna('N')
    tem = pd.crosstab(pv2.vindall, pv2.vacant)
    tp.append(tem[1]['Y'])
    tn.append(tem[0]['N'])
    fp.append(tem[0]['Y'])
    fn.append(tem[1]['N'])
    
# # import matplotlib.dates as mdates
# fig, ax = plt.subplots()

# plt.plot(dates, tp)
# # plt.plot(dates, tn)
# plt.plot(dates, fp)
# plt.plot(dates, fn);

# x1,x2,y1,y2 = plt.axis()
# plt.axis([x1,x2,0,y2]);
# plt.legend(['Both predicted vacant','Vacant (only postal)','Vacant (only survey)'],'lower right');

# fig.autofmt_xdate()

# # fig.savefig(path+'/outputs/pv.png')

pv2.head()

pv2 = pd.merge(tci, pv[(pv.date==dt.datetime(2015,10,1))], how='left', left_on='parcel', right_on='PARCEL')
pv2['vindall'] = pv2['vindall'].fillna('N')
pd.crosstab(pv2.vindall, pv2.vacant)

pv2 = pd.melt(pv.pivot(index='PARCEL',columns='date').fillna('N').reset_index(),id_vars='PARCEL',value_name='vindall')
pv2.head()

def fun(x):
    total = 0
#     print(dates[x.parcel])
    recent = pv_dates[pv_dates<dates[x.PARCEL]
                      
    for i,j in enumerate(x[recent]):
        if x[recent][-i-1] == 'Y':
            total += 1
        else:
            return total
    return total

pv_dates = pv.groupby('date').last().index
# pv2 = pv.pivot(index='PARCEL',columns='date').fillna('N').reset_index()
# pv2.columns = pv2.columns.droplevel()
# pv2.columns.values[0] = 'parcel'
# pv2['pv_count'] = pv2.apply(fun, axis=1)
# pv2 = pv2.set_index('parcel')
# pv2
pv_dates

tci = pd.merge(tci, pv2[['pv_count']], how='left', left_on='parcel',right_index=True)
tci['pv_count'] = tci['pv_count'].fillna(0)

# tci.loc[tci.pv_count>0,'pv_count2'] = 1
# tci.loc[tci.pv_count==0,'pv_count2'] = 0

# tci.loc[tci.delq_total_ratio>0.7,'del_ratio'] = 1
# tci.loc[tci.delq_total_ratio<=0.7,'del_ratio'] = 0

# t = tci[(tci.pv_count2==1) & (tci.del_ratio==1)].vacant
# print(sum(t),len(t))

# t = tci[(tci.pv_count2==0) & (tci.del_ratio==1)].vacant
# print(sum(t),len(t))

# t = tci[(tci.pv_count2==1) & (tci.del_ratio==0)].vacant
# print(sum(t),len(t))

# t = tci[(tci.pv_count2==0) & (tci.del_ratio==0)].vacant
# print(sum(t),len(t))


cols = list(tci.columns)
cols.remove('vacant')
cols.remove('Date')

tci[cols].to_csv(path+'/data/model_data/tci_2_1.csv', index=False)

tci.columns

tci.info()
