==> notebooks/2.3 Merge data [foreclosures, sheriff's auction, transfers, sales].ipynb <==
import matplotlib.pyplot as plt
import datetime as dt
import seaborn as sns
import pandas as pd
import numpy as np
import os

%matplotlib inline

path = 'Z:/largetransfer/luc/carter'
# print(path)

tci = pd.read_csv(path+'/data/model_data/tci_1_0.csv', parse_dates=['Date'], dtype={'Parcel ID':str})
cols = ['parcel','Date','vacant']
tci = tci[cols]

ppns = set(tci['parcel'])
dates = dict(zip(tci.parcel, tci.Date))

print(tci.shape)

## Foreclosure data

1. Collects observation in the `foreclosure_filings2.csv` according to dates 2015 and below.
2. Sorts based on dates
3. Takes recent abservation based on variable, "case no",(i.e recent case numbers)

fc = pd.read_csv(path+'/data/clean_data/foreclosure_filings2.csv', parse_dates=[2])
fc = fc.loc[fc.apply(lambda x: x.filedate < dates[x.parcel], axis=1)]

fc.head()

max(fc.filedate)

fc.info()

fc.head(1)
fc = fc.sort_values('filedate')
fc = fc.groupby('caseno').last()

fc['pamount'] = fc['pamount'].apply(lambda x: float(x[1:].replace(',','')))

#### Active foreclosure
1. Merges parcels that are under active foreclosure

tem = fc.loc[fc.status=='Active',['parcel']]
tem['active_fc'] = 1
tem = tem.drop_duplicates()
tem = tem.set_index('parcel')

tci = pd.merge(tci, tem, \
             how='left', left_on='parcel', right_index=True)
print(tci.shape)

tem = fc[['parcel','pamount']].groupby('parcel').last()
tem.columns = ['foreclosure_flag']

tci = pd.merge(tci, tem, \
             how='left', left_on='ppn', right_index=True)

fc_copy = fc.reset_index()

fc_copy = fc_copy[['parcel','filedate']].groupby('parcel').count().reset_index()

fc_copy['fc_1yr'] = fc_copy.apply(lambda x: \
        len(fc[(fc.parcel == x.parcel) & (fc.filedate > dates[x.parcel]-pd.DateOffset(years=1))]), axis=1)

fc_copy = fc_copy.set_index('parcel')

tci = pd.merge(tci, fc_copy[['fc_1yr']], \
               how='left', left_on='parcel', right_index=True)
tci.loc[tci.fc_1yr.isnull(), 'fc_1yr'] = 0

# fc_copy = fc.reset_index()

# fc_copy = fc_copy[['parcel','filedate']].groupby('parcel').count().reset_index()

# fc_copy['fc_1yr'] = fc_copy.apply(lambda x: \
#         len(fc[(fc.parcel == x.parcel) & (fc.filedate > dates[x.parcel]-pd.DateOffset(years=1))]), axis=1)


fc_copy = fc.reset_index()

fc_copy = fc_copy[['parcel','filedate']].groupby('parcel').count().reset_index()

fc_copy['fc_2yr'] = fc_copy.apply(lambda x: \
        len(fc[(fc.parcel == x.parcel) & (fc.filedate > dates[x.parcel]-pd.DateOffset(years=2))]), axis=1)

fc_copy = fc_copy.set_index('parcel')

tci = pd.merge(tci, fc_copy[['fc_2yr']], \
               how='left', left_on='parcel', right_index=True)
tci.loc[tci.fc_2yr.isnull(), 'fc_2yr'] = 0

# fc_count = fc[fc.filedate > dt.datetime(2013,6,1)].groupby('parcel').count()[['status']]
# fc_count = fc_count.rename(columns={'status':'fc_2yr'})
# tci = pd.merge(tci, fc_count, left_on='ppn',right_index=True, how='left')
# tci.loc[tci.fc_2yr.isnull(),'fc_2yr'] = 0

recent_fc = fc.sort_values('filedate').groupby('parcel').last().reset_index()
fc_days = pd.DataFrame(recent_fc[['parcel','filedate']])
fc_days['fc_days'] = fc_days.apply(lambda x: (dates[x.parcel]-x.filedate).days,axis=1)
fc_days = fc_days.set_index('parcel')
tci = pd.merge(tci, fc_days[['fc_days']], left_on='parcel',right_index=True, how='left')
tci.loc[tci.fc_days.isnull(),'fc_days'] = max(tci[tci.fc_days.notnull()].fc_days)+1

# tci.loc[(tci.ppn.isin(set(fc[(fc.filedate < dt.datetime(2007,6,1)) & ((fc.ldispos=='DIS. W/O PREJ'))].parcel))),'fc_dis_w/o_prej'] = 1
# tci.loc[~(tci.ppn.isin(set(fc[(fc.filedate < dt.datetime(2007,6,1)) & ((fc.ldispos=='DIS. W/O PREJ'))].parcel))),'fc_dis_w/o_prej'] = 0
# pd.crosstab(tci['fc_dis_w/o_prej'], tci.vacant)

# tci.loc[(tci.ppn.isin(set(fc[(fc.ldispos=='DIS. W/O PREJ')].parcel))),'fc_dis_w/o_prej'] = 1
# tci.loc[(~tci.ppn.isin(set(fc[(fc.ldispos=='DIS. W/O PREJ')].parcel))),'fc_dis_w/o_prej'] = 0
# pd.crosstab(tci['fc_dis_w/o_prej'], tci.vacant)

# tci.loc[(tci.ppn.isin(set(fc[(fc.filedate > dt.datetime(2014,6,1)) & (fc.ldispos=='NEWLY FILED')].parcel))),'fc_newly_filed'] = 1
# tci.loc[~(tci.ppn.isin(set(fc[(fc.filedate > dt.datetime(2014,6,1)) & (fc.ldispos=='NEWLY FILED')].parcel))),'fc_newly_filed'] = 0
# pd.crosstab(tci['fc_newly_filed'], tci.vacant)

# for dis1,dis2 in [['NEWLY FILED','fc_newly_filed'], ['DEFAULT','fc_default'], ['DISP.OTHER','fc_disp_other'], ['DIS. W/O PREJ','fc_dis_wo_prej']]:
#     recent_fc = fc[(fc.ldispos==dis1)].sort_values('filedate').groupby('parcel').last()
#     fc_days = pd.DataFrame(recent_fc.apply(lambda x: (dt.datetime(2015,6,1)-x.filedate).days,axis=1),columns=[dis2])
#     tci = pd.merge(tci, fc_days, left_on='ppn',right_index=True, how='left')
#     tci.loc[tci[dis2].isnull(),dis2] = max(tci[dis2])
# print tci.shape

## Sheriff's auction

1. Create a new column for sherif sales days adds column, 'sa_days' to `tci_1_0.csv`

sa = pd.read_csv(path+'/data/clean_data/sheriff_auction.csv', parse_dates=[2], encoding="ISO-8859-1")
sa = sa.loc[sa.apply(lambda x: x.salesdt < dates[x.parcel], axis=1)]

sa.columns

recent_sa = sa.sort_values('salesdt').groupby('parcel').last().reset_index()
sa_days = pd.DataFrame(recent_sa[['parcel','salesdt']])
sa_days['sa_days'] = sa_days.apply(lambda x: (dates[x.parcel]-x.salesdt).days,axis=1)
sa_days = sa_days.set_index('parcel')
tci = pd.merge(tci, sa_days[['sa_days']], left_on='parcel',right_index=True, how='left')
tci.loc[tci.sa_days.isnull(),'sa_days'] = max(tci[tci.sa_days.notnull()].fc_days)+1

# sa = sa.set_index('parcel')

sa['days'] = pd.DataFrame(sa.apply(lambda x: (dt.datetime(2015,6,1)-x.salesdt).days,axis=1),columns=['sa_days'])

sa.loc[sa.minbid.notnull(),'minbid'] = sa[sa.minbid.notnull()].minbid.apply(lambda x: float(x[1:].replace(',','')))

# for status,name in [['NO BID SOLD TO LAND BANK','sa_no_bid_lb'],['NO BID FORFEIT TO STATE','sa_no_bid_state'] \
#                     ,['ACTIVE','sa_active'],['SOLD','sa_sold']]:
#     tci = pd.merge(tci, sa.loc[sa.status == status,['days']], left_on='ppn', right_index=True, how='left')
#     tci = tci.rename(columns={'days':name})
# print(tci.shape)

sa['year'] = sa.salesdt.apply(lambda x: x.year)

def parse_amount(x):
    return float(str(x)[1:].replace(',',''))

# sa.loc[sa.appraisal.notnull(), 'appraisal'] = sa.loc[sa.appraisal.notnull(), 'appraisal'].apply(parse_amount)

# sa['ssold'] = sa.ssold.apply(lambda x: 1 if x=='Yes' else 0)

# tem = sa.loc[sa.salesdt>dt.datetime(2013, 1, 1),['parcel','ssold']].groupby('parcel').sum()
# tem.columns = ['auction_13']
# tem.loc[tem['auction_13'].isnull(),'auction_13'] = 0
# tci = pd.merge(tci, tem, \
#               how='left', left_on='ppn', right_index=True)
# tci.loc[tci['auction_13'].isnull(),'auction_13'] = 0
# tci.loc[(tci.vacant>-1),['auction_13','vacant']].groupby('auction_13').agg([sum,len])

# sum(tci.auction_13)

# tem = sa[['parcel','ssold']].groupby('parcel').sum()
# tem.columns = ['auction_all']
# tem.loc[tem['auction_all'].isnull(),'auction_all'] = 0
# tci = pd.merge(tci, tem, \
#               how='left', left_on='ppn', right_index=True)
# tci.loc[tci['auction_all'].isnull(),'auction_all'] = 0

# tci.loc[(tci.vacant>-1),['auction_all','vacant']].groupby('auction_all').agg([sum,len])

# tci[tci.auction_13.isnull()] = 0
# tci[tci.auction_13 > 0] = 1
# tci[tci.auction_all.isnull()] = 0
# tci[tci.auction_all > 0] = 1

del sa
del fc

## Transfers
1. Adds a column to `tci_1_0.csv` for column columns indicating days since transfer `mdate`
2. Adds a column to `tci_1_0.csv` for column columns indicating days since sheriff's `t_shf_deed`


t = pd.read_csv(path+'/data/clean_data/transfers.csv', parse_dates=['mdate'], dtype=str)
t = t.sort_values('mdate', ascending=False)

t = t.loc[t.apply(lambda x: x.mdate < dates[x.PROPERTY_NUMBER], axis=1)]
# t.info()

# days since transfer

days_since_transfer = t.groupby('PROPERTY_NUMBER').first()
days_since_transfer = pd.merge(days_since_transfer[['mdate']], tci[['parcel','Date']], how='left', left_index=True, right_on='parcel')

tci = pd.merge(tci, days_since_transfer[['mdate']], left_on='parcel', right_index=True,how='left')
tci.loc[tci.mdate.isnull(),'mdate'] = dt.datetime(2006,1,1)
tci['mdate'] = tci.apply(lambda x: (x.Date - x.mdate).days, axis=1)
# days_since_transfer = days_since_transfer.set_index('parcel')
tci = tci.rename(columns={'mdate':'t_days'})
# tci.loc[tci.t_days.isnull(),'t_days'] = max(tci.t_days)

# days since sheriff's deed type

recent_t = t[(t.mdate > dt.datetime(2006,1,1)) & (t.DEED_TYPE=='Sheriffs Deed')].groupby('PROPERTY_NUMBER').first()
recent_t = pd.merge(recent_t, tci[['parcel','Date']], how='left', left_index=True, right_on='parcel')
recent_t['t_shf_deed'] = recent_t.apply(lambda x: (x.Date-x.mdate).days, axis=1)
recent_t = recent_t.set_index('parcel')

tci = pd.merge(tci, recent_t[['t_shf_deed']], left_on='parcel',right_index=True, how='left')
tci.loc[tci.t_shf_deed.isnull(),'t_shf_deed'] = max(tci[tci.t_shf_deed.notnull()].t_shf_deed)+1

# number of transfers per year since 2006

tci = pd.merge(tci, t.loc[t.mdate > dt.datetime(2006,1,1), ['PROPERTY_NUMBER','mdate']] \
               .groupby('PROPERTY_NUMBER').count(), how='left', left_on='parcel', right_index=True)
tci['mdate'] = tci['mdate'].fillna(0)
tci = tci.rename(columns = {'mdate':'t_num'})
tci['t_num'] = tci.apply(lambda x: x.t_num/((x.Date - dt.datetime(2006,1,1)).days/365.0), axis = 1)

# tci = pd.merge(tci, t[t.mdate > dt.datetime(2013, 6, 1)].groupby('PROPERTY_NUMBER').count()[['mdate']], how='left', left_on='ppn',right_index=True)
# tci = tci.rename(columns={'mdate':'num_of_transfers'})

# deeds = list(t.groupby('DEED_TYPE').count().sort_values('mdate',ascending=False).index[0:5])
# tci = pd.merge(tci, t[['DEED_TYPE', 'PROPERTY_NUMBER']].groupby('PROPERTY_NUMBER').first(), \
#          how='left', left_on='ppn',right_index=True)
# tci.DEED_TYPE = tci.DEED_TYPE.apply(lambda x: x if x in deeds else 'Other')

## Armslength sales


al = pd.read_csv(path+'/data/clean_data/armslength.csv', dtype=str)
al = al.drop_duplicates()

al.columns

# al = pd.read_csv(path+'/data/clean_data/armslength.csv', dtype=str)
# al = al.drop_duplicates()

months = dict(zip(['JAN','FEB','MAR','APR','MAY','JUN','JUL','AUG','SEP','OCT','NOV','DEC'],range(1,13)))

def parse_date(x):
    day = int(x[0:2])
    month = months[x[2:5]]
    year = int(x[5:])
    return dt.datetime(year,month,day)

al['date'] = al.mdate.apply(parse_date)
al = al.sort_values('date')
# al = al[al.date < dt.date(2015,6,1)]

al = al.loc[al.apply(lambda x: x.date < dates[x.PROPERTY_NUMBER], axis=1)]

# number of sales per year since 2006

tci = pd.merge(tci, al.loc[al.date > dt.datetime(2006,1,1),['date','PROPERTY_NUMBER']]\
               .groupby('PROPERTY_NUMBER')\
               .agg(lambda x: len(x)\
                    /((dates[x.PROPERTY_NUMBER.iloc[0]] - dt.datetime(2006,1,1)).days/365.0))\
                    ,how='left', left_on='parcel', right_index=True)
tci = tci.rename(columns = {'date':'al_num'})
tci.loc[tci.al_num.isnull(),'al_num'] = 0

# days since most recent sale since 2006

tci = pd.merge(tci, al.loc[al.date > dt.datetime(2006,1,1),['date','PROPERTY_NUMBER']]\
               .groupby('PROPERTY_NUMBER')\
               .last()\
                    ,how='left', left_on='parcel', right_index=True)
tci = tci.rename(columns = {'date':'al_date'})

tci.loc[tci.al_date.isnull(),'al_date'] = dt.datetime(2006,1,1)
tci['al_days'] = tci.apply(lambda x: (x.Date-x.al_date).days, axis=1)


tci.shape

cols = list(tci.columns)
cols.remove('vacant')
cols.remove('Date')
print(cols)

tci = tci.fillna(0)

tci[cols].to_csv(path+'/data/model_data/tci_2_3.csv',index=False)

tci[cols].info()
