==> notebooks/3.0 Combine data.ipynb <==
# 3.0 Combine data

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns
import math
from sklearn.preprocessing import Imputer
from sklearn.base import TransformerMixin
from sklearn import cross_validation
from sklearn.grid_search import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn import preprocessing
from sklearn.cross_validation import cross_val_score
import os

def sound():
    os.system('afplay /System/Library/Sounds/Sosumi.aiff')

%matplotlib inline
sns.set_context('notebook')

path = '/'.join(os.getcwd().split('/')[:-1])
print(path)

### Combines the data sets:
1. `tci_2_0.csv`
2. `tci_2_1.csv`
3. `tci_2_2.csv`
3. `tci_2_3.csv`

df0 = pd.read_csv(path+'/data/model_data/tci_2_0.csv', dtype={'Parcel ID':str}, index_col='parcel')
df1 = pd.read_csv(path+'/data/model_data/tci_2_1.csv', dtype={'Parcel ID':str}, index_col='parcel')
df2 = pd.read_csv(path+'/data/model_data/tci_2_2.csv', dtype={'Parcel ID':str}, index_col='parcel')
df3 = pd.read_csv(path+'/data/model_data/tci_2_3.csv', dtype={'Parcel ID':str}, index_col='parcel')

df = pd.merge(df0, df1[[col for col in df1.columns if col not in ['Date','vacant']]], how='inner', left_index=True, right_index=True)
df = pd.merge(df, df2[[col for col in df2.columns if col not in ['Date','vacant']]], how='inner', left_index=True, right_index=True)
df = pd.merge(df, df3[[col for col in df3.columns if col != 'Date']], how='inner', left_index=True, right_index=True)

df.columns

df.loc[df.occup.notnull(),'occup'] = df[df.occup.notnull()].occup.apply(lambda x: int(x[0]))

df.loc[df.occup.isnull(),'occup'] = 1

tem = df[['SPA_NAME','totusabl']].groupby('SPA_NAME').agg(np.median)
tem = dict(tem['totusabl'])
df.loc[df.totusabl==0, 'totusabl'] = df[df.totusabl==0].apply(lambda x: tem[x.SPA_NAME], axis=1)

df['price_per_sqft'] = df['tmktval']/df['totusabl']

# tem = df[['SPA_NAME','tmktval','condition_value14','GRAND_TOTAL_BALANCE','sqft']].groupby('SPA_NAME').agg([np.median,np.mean])
# tem

# val = dict(tem['condition_value14']['mean'])
# df['cond_diff'] = df.apply(lambda x: x.condition_value14 - val[x.SPA_NAME], axis=1)

# val = dict(tem['tmktval']['median'])
# df['tmktval_diff'] = df.apply(lambda x: x.tmktval - val[x.SPA_NAME], axis=1)

# val = dict(tem['GRAND_TOTAL_BALANCE']['median'])
# df['balance_diff'] = df.apply(lambda x: x.GRAND_TOTAL_BALANCE - val[x.SPA_NAME], axis=1)

# k = 0.00005
# b = np.linspace(0,20000,20)
# sns.distplot(df.loc[df.vacant==1,'GRAND_TOTAL_BALANCE'], bins=b,kde_kws={'bw':k})
# sns.distplot(df.loc[df.vacant==0,'GRAND_TOTAL_BALANCE'], bins=b,kde_kws={'bw':k})
# plt.axis([min(b),max(b),0,.0009]);

# k = 0.00005
# b = np.linspace(-500,20000,20)
# sns.distplot(df.loc[df.vacant==1,'balance_diff'], bins=b,kde_kws={'bw':k})
# sns.distplot(df.loc[df.vacant==0,'balance_diff'], bins=b,kde_kws={'bw':k})
# plt.axis([min(b),max(b),0,.0009]);

# b = np.linspace(0,150000,40)
# sns.distplot(df.loc[df.vacant==1,'tmktval'], bins=b)
# sns.distplot(df.loc[df.vacant==0,'tmktval'], bins=b)
# plt.axis([0,max(b),0,0.000025]);

# sns.distplot(df.loc[df.vacant==1,'sqft'], bins=np.linspace(0,160,40))
# sns.distplot(df.loc[df.vacant==0,'sqft'], bins=np.linspace(0,160,40))
# plt.axis([0,150,0,0.035]);

# k = 0.5
# b = np.linspace(0,10,11)
# sns.distplot(df.loc[df.vacant==1,'condition_value14'], bins=b,kde_kws={'bw':k})
# sns.distplot(df.loc[df.vacant==0,'condition_value14'], bins=b,kde_kws={'bw':k})
# plt.axis([0,10,0,.9]);

# k = 3000
# b = np.linspace(-50000,50000,50)
# sns.distplot(df.loc[df.vacant==1,'tmktval_diff'], bins=b,kde_kws={'bw':k})
# sns.distplot(df.loc[df.vacant==0,'tmktval_diff'], bins=b,kde_kws={'bw':k})
# plt.axis([-50000,50000,0,.00005]);

# k = .3
# b = np.linspace(-5,5,20)
# sns.distplot(df.loc[df.vacant==1,'cond_diff'], bins=b,kde_kws={'bw':k})
# sns.distplot(df.loc[df.vacant==0,'cond_diff'], bins=b,kde_kws={'bw':k})
# plt.axis([-5,5,0,.9]);

# val = dict(tem['tmktval']['median'])
# df['mktval_diff'] = df.apply(lambda x: x.tmktval - val[x.SPA_NAME], axis=1)

# cols_temp = df.columns[:-4]

# df['missing_values'] = df.apply(lambda x: len(x[x.isnull()]),axis=1)
to_remove = ['Parcel ID','parcel','Date','Survey Category','Survey Grade',\
             'condition13', 'condition14', 'al_date','SPA_NAME','NAME10','style_filtered','rextwall','LENDER_PROCESS_TYPE']

features = list(df.columns.values)
for i in to_remove:
    if i in features:
        features.remove(i)

for i,j in enumerate(features):
    print(i,j)

# from sklearn import tree

# clf = tree.DecisionTreeClassifier(max_depth=2,min_samples_leaf=5)
# # iris = load_iris()
# clf = clf.fit(df3[features], df3.vacant)
# tree.export_graphviz(clf, out_file=path+'/tree.dot',feature_names=features, impurity=False, label=None,filled=True,class_names=['Not vacant','Vacant']) 

print(df[features].columns[df[features].dtypes == 'object'])

# transform categorical variables (including NaN's)
# objs = df[features].columns[df[features].dtypes == 'object']
# prefixes = ['style','ext','lender']

# objs = ['LENDER_PROCESS_TYPE']
# prefixes = ['lender']

# df2 = df[objs]
# df3 = pd.concat([pd.get_dummies(df2[col], prefix=prefixes[index]) for index, col in enumerate(df2)], axis=1)

# for i in objs:
#     if i in features:
#         features.remove(i)
df3 = pd.DataFrame()
df3[features] = df[features]
df3 = df3.fillna(0)

print(df3.shape)

df3.to_csv(path+'/data/model_data/tci_final.csv')
